+++ ichnaea/api/submit/submit_v2/tests.py
@@ -13,7 +13,7 @@ class TestView(BaseSubmitTest, CeleryAppTestCase):
 
     url = '/v1/geosubmit'
     metric = 'geosubmit'
-    metric_url = 'request.v1.geosubmit'
+    metric_path = 'path:v1.geosubmit'
     status = 200
     radio_id = 'radioType'
     cells_id = 'cellTowers'