+++ ichnaea/api/submit/submit_v3/tests.py
@@ -13,7 +13,7 @@ class TestView(BaseSubmitTest, CeleryAppTestCase):
 
     url = '/v2/geosubmit'
     metric = 'geosubmit2'
-    metric_url = 'request.v2.geosubmit'
+    metric_path = 'path:v2.geosubmit'
     status = 200
     radio_id = 'radioType'
     cells_id = 'cellTowers'